Can now remove a pantry and It will update other users who use that pantry.

+++ lib/pages/pantry.dart
@@ -23,6 +23,7 @@ class _PantryState extends State<Pantry> {
   Map<String, DocumentReference>
       _pantryMap; // private NOTE: bad design - it will fuck with users collaborating on multiple pantries with the same name
   bool loading = true;
+  bool isOwner = false;
 
   @override
   void initState() {
@@ -41,6 +42,9 @@ class _PantryState extends State<Pantry> {
       // go through each doc ref and add to list of pantry names + map
       String pantryName = "";
       await ref.get().then((DocumentSnapshot snapshot) {
+        if(snapshot.data()['Owner'].id == widget.user.uid){
+          isOwner = true;
+        }
         if (ref == user.PPID) {
           pantryName = snapshot.data()['Name'] + "*";
         } else {
@@ -203,46 +207,47 @@ class _PantryState extends State<Pantry> {
 
    Future<void> deletePantry(DocumentReference doc) async {
     // delete pantry from list of pantries
-    await doc
-        .delete()
-        .then((value) => print("SUCCESS: $doc has been deleted from pantries"))
-        .catchError((error) =>
-            print("FAILURE: couldn't delete $doc from pantries: $error"));
-    // delete pantry from user pantries
-    await firestoreInstance
-        .collection('users')
-        .doc(user.uid)
-        .update({
-          'PantryIDs': FieldValue.arrayRemove([_selectedPantry])
-        })
-        .then((value) =>
-            print("SUCCESS: $doc has been deleted from user pantries"))
-        .catchError((error) =>
-            print("FAILURE: couldn't delete $doc from user pantries: $error"));
-    // if pantry is primary, remove it from PPID
-    if (doc == user.PPID) {
-      await firestoreInstance
-          .collection('users')
-          .doc(user.uid)
-          .update({'PPID': FieldValue.delete()})
-          .then((value) =>
-              print("SUCCESS: $doc has been deleted from user pantries"))
-          .catchError((error) => print(
-              "FAILURE: couldn't delete $doc from user pantries: $error"));
-      if (_pantryMap.isNotEmpty) {
-        var entryList = _pantryMap.entries.toList();
-        FirebaseFirestore.instance
-            .collection("users")
-            .doc(user.uid)
-            .update({
-          'PPID': entryList[0].value,
-        }).catchError((error) =>
-            print("Failed to set new primary pantry: $error"));
-          updatePantries(entryList[0].key, true, true);
-      }
-    }
+     if(isOwner) {
+       var snap = await doc.get();
+       List altUsers = snap.data()['AltUsers'];
+       altUsers.forEach((element) {removePantryFromUser(element.id, doc);});
+       await doc
+           .delete()
+           .then((value) =>
+           print("SUCCESS: $doc has been deleted from pantries"))
+           .catchError((error) =>
+           print("FAILURE: couldn't delete $doc from pantries: $error"));
+     }
+     removePantryFromUser(user.uid, doc);// delete pantry from user pantries
   }
 
+  removePantryFromUser(id, doc) async{
+
+    // if pantry is primary, remove it from PPID
+    var tempUser = await firestoreInstance
+        .collection('users')
+        .doc(id).get();
+    var tempPPID = tempUser.data()['PPID'];
+    if (doc == tempUser.data()['PPID']) {
+      tempPPID = null;
+      for(int i = 0; i< tempUser.data()['PantryIDs'].length; i++){
+        if(tempUser.data()['PantryIDs'][i] != doc){
+          tempPPID = tempUser.data()['PantryIDs'][i];
+          break;
+        }
+    }
+      }
+    await firestoreInstance
+        .collection('users')
+        .doc(id)
+        .update({
+      'PPID': tempPPID,
+      'PantryIDs': FieldValue.arrayRemove([_selectedPantry])
+    }).then((value) =>
+        print("SUCCESS: $doc has been deleted from user pantries"))
+        .catchError((error) =>
+        print("FAILURE: couldn't delete $doc from user pantries: $error"));
+    }
   showDeleteDialog(
       BuildContext context, String pantryName, DocumentReference doc) {
     Widget cancelButton = TextButton(
